+++ cmd/due.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"os"
 	"punchlist/task"
 	"strconv"
 	"strings"
@@ -21,7 +22,7 @@ func newDueCmd() *cobra.Command {
 			// parse inputs
 			id, err := strconv.Atoi(args[0])
 			if err != nil {
-				fmt.Printf("Invalid task ID: %v\n", err)
+				fmt.Fprintf(os.Stderr, "Invalid task ID: %v\n", err)
 				return
 			}
 
@@ -29,7 +30,7 @@ func newDueCmd() *cobra.Command {
 			dueInput := strings.Join(args[1:], " ")
 			dueTime, err := parseDue(dueInput)
 			if err != nil {
-				fmt.Printf("Invalid due date: %v\n", err)
+				fmt.Fprintf(os.Stderr, "Invalid due date: %v\n", err)
 				return
 			}
 
@@ -38,13 +39,13 @@ func newDueCmd() *cobra.Command {
 				if printNotPunchlistError(err) {
 					return
 				}
-				fmt.Printf("Error finding task: %v\n", err)
+				fmt.Fprintf(os.Stderr, "Error finding task: %v\n", err)
 				return
 			}
 
 			t, err := task.Parse(taskPath)
 			if err != nil {
-				fmt.Printf("Error parsing task: %v\n", err)
+				fmt.Fprintf(os.Stderr, "Error parsing task: %v\n", err)
 				return
 			}
 
@@ -73,7 +74,7 @@ func newDueCmd() *cobra.Command {
 			t.Body = joinBlocks(pre, logSection)
 
 			if err := t.Write(taskPath); err != nil {
-				fmt.Printf("Error updating task: %v\n", err)
+				fmt.Fprintf(os.Stderr, "Error updating task: %v\n", err)
 				return
 			}